backend/requirements.txt
asgiref==3.10.0
boto3==1.29.7
botocore==1.32.7
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.3.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
jmespath==1.0.1
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2025.2
s3transfer==0.8.0
sendgrid==6.11.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
urllib3==2.0.7
gunicorn
django-storages